Pulse Generator and Seconds Counter
Full_Pulse outputs a variable speed pulse with a 50% duty ratio. Two select bit choose between 32, 64, 128, or modulated pulse rate. The modulated pulse length is based on the time elapsed after it is selected. The Seconds module outputs a pulse every second. Both modules relay on the 100Mhz clock available on the Bayes 3.