Fully Tested Pulse Generator
This machine has been placed on the FPGA board with flying colors. There is a routing error that can be voided with no impact to any clock speeds that I could measure with a stopwatch. Timing needs to be verified with entire system.
I/O
start: Allows time to progress and enables clock on posedge
reset: Freezes pulse and forces time back to on next startup
pulseBase : This is the base clock assumed to be at 100Mz
select: Two bit mode select
pulseOut: desired pulse based on select and clock